+++ src/azure-cli/HISTORY.rst
@@ -62,7 +62,7 @@ Release History
 
 **SQL**
 
-* New commands `sql db sensitivity-classification show/list/update/delete` and `sql db sensitivity-classification recommendation list/enable/disable` to manage sensitivity classifications for SQL databases.
+* New commands `sql db classification show/list/update/delete` and `sql db classification recommendation list/enable/disable` to manage sensitivity classifications for SQL databases.
 
 2.0.78
 ++++++